requirements.txt
tensorflow==2.3.1
Pillow
numpy
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-restful
flask_cors
flask_api
git+https://github.com/mitsuhiko/flask-oauth
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauth2==1.9.0.post1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
werkzeug==0.16.1
gunicorn
configparser
opencv-python==4.4.0.46
db-sqlite3
